requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
ecdsa==0.13.3
Flask==1.1.1
Flask-QRcode==3.0.0
flaskwebgui==0.0.8
hidapi==0.7.99.post21
hwi==1.0.3
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
libusb1==1.7.1
MarkupSafe==1.1.1
mnemonic==0.18
more-itertools==7.2.0
packaging==19.2
pbkdf2==1.3
Pillow==6.2.0
pluggy==0.13.1
psutil==5.6.7
py==1.8.0
pyaes==1.6.1
pyparsing==2.4.5
pytest==5.2.2
python-dotenv==0.10.3
qrcode==6.1
requests==2.22.0
six==1.12.0
stem==1.7.1
typing-extensions==3.7.4
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.16.0
whichcraft==0.6.1
zipp==0.6.0
